+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Actions/AddLabelsToCardAction.cs
@@ -2,6 +2,9 @@
 using System.Linq;
 using System.Threading.Tasks;
 using TrelloDotNet.AutomationEngine.Interface;
+using TrelloDotNet.Control;
+using TrelloDotNet.Model;
+using TrelloDotNet.Model.Options;
 using TrelloDotNet.Model.Webhook;
 
 namespace TrelloDotNet.AutomationEngine.Model.Actions
@@ -77,7 +80,10 @@ namespace TrelloDotNet.AutomationEngine.Model.Actions
 
             if (updateNeeded)
             {
-                await trelloClient.UpdateCardAsync(card);
+                await trelloClient.UpdateCardAsync(card.Id, new List<QueryParameter>()
+                {
+                    new QueryParameter(CardFieldsType.LabelIds.GetJsonPropertyName(), card.LabelIds)
+                });
                 processingResult.AddToLog($"Added labels '{string.Join(",", LabelIds)}' to card '{webhookAction.Data.Card.Name}'");
                 processingResult.ActionsExecuted++;
             }