+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Actions/AddMembersToCardAction.cs
@@ -2,6 +2,9 @@
 using System.Linq;
 using System.Threading.Tasks;
 using TrelloDotNet.AutomationEngine.Interface;
+using TrelloDotNet.Control;
+using TrelloDotNet.Model.Options;
+using TrelloDotNet.Model;
 using TrelloDotNet.Model.Webhook;
 
 namespace TrelloDotNet.AutomationEngine.Model.Actions
@@ -75,7 +78,10 @@ namespace TrelloDotNet.AutomationEngine.Model.Actions
 
             if (updateNeeded)
             {
-                await trelloClient.UpdateCardAsync(card);
+                await trelloClient.UpdateCardAsync(card.Id, new List<QueryParameter>()
+                {
+                    new QueryParameter(CardFieldsType.MemberIds.GetJsonPropertyName(), card.MemberIds)
+                });
                 processingResult.AddToLog($"Added members '{string.Join(",", MemberIds)}' to card '{webhookAction.Data.Card.Name}'");
                 processingResult.ActionsExecuted++;
             }