+++ TrelloDotNet/TrelloDotNet/AutomationEngine/Model/Actions/SetCardStartFieldValue.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using TrelloDotNet.Control;
 using TrelloDotNet.Model;
+using TrelloDotNet.Model.Options;
 
 namespace TrelloDotNet.AutomationEngine.Model.Actions
 {
@@ -41,6 +43,12 @@ namespace TrelloDotNet.AutomationEngine.Model.Actions
                     throw new ArgumentOutOfRangeException();
             }
             return false;
+        }
+
+        /// <inheritdoc />
+        public QueryParameter GetQueryParameter(Card card)
+        {
+            return SetIfNeeded(card) ? new QueryParameter(CardFieldsType.Start.GetJsonPropertyName(), card.Start) : null;
         }
 
         /// <summary>